Exception handling

+++ src/main/java/com/valneirego/commerce/services/ProductService.java
@@ -4,10 +4,15 @@ package com.valneirego.commerce.services;
 import com.valneirego.commerce.dto.ProductDTO;
 import com.valneirego.commerce.entities.Product;
 import com.valneirego.commerce.repositories.ProductRepository;
+import com.valneirego.commerce.services.exeptions.DatabaseException;
+import com.valneirego.commerce.services.exeptions.ResouceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 
@@ -19,7 +24,7 @@ public class ProductService {
     @Transactional(readOnly = true)
     public ProductDTO findByID(Long id) {
 
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).orElseThrow(() -> new ResouceNotFoundException("Recurso não encontrado"));
         return new ProductDTO(product);
     }
 
@@ -43,18 +48,38 @@ public class ProductService {
 
 
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product entity = repository.getReferenceById(id);
 
-        copyDtoToEntity(dto, entity);
+        try {
 
-        repository.save(entity);
+            Product entity = repository.getReferenceById(id);
 
-        return new ProductDTO(entity);
+            copyDtoToEntity(dto, entity);
+
+            repository.save(entity);
+
+            return new ProductDTO(entity);
+        }
+
+        catch (EntityNotFoundException e){
+            throw new ResouceNotFoundException("Recurso não encontrado");
+        }
+
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete (Long id) {
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResouceNotFoundException("Recurso não encontrado");
+        }
+
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de intgidade refereincial");
+        }
+
+
     }
 
     private static void copyDtoToEntity(ProductDTO dto, Product entity) {
